RFC 10: Default C flags
Habitat's core plans should provide a reliable base system on which
Habitat users can build applications. To meet this goal, core-plans
software written in the C programming language should use compiler and
linker flags that provide a level of performance and security at least
as good as other Linux distributions.

Multiple users have attempted to make progress on this issue in the
past. For example,

- https://github.com/habitat-sh/core-plans/issues/587
- https://github.com/habitat-sh/core-plans/pull/881

and the lack of optimization was recently discussed in the Habitat
Slack.

NOTE FOR REVIEW: I took a "slow and steady" approach to implementing
this on the assumption that this would be mostly an effort that
happens outside the core Hab team. I think a faster path is possible.

Signed-off-by: Steven Danna <steve@chef.io>